Use the 'test-tools' package for the integration tests.
This removes all Kubernetes and KUDO specific code in favor of the functionality provided by 'test-tools'.
Cassandra specific functions are in the local 'cassandra' package.
